+++ src/hexdoc/jinja/__init__.py
@@ -2,9 +2,16 @@ __all__ = [
     "IncludeRawExtension",
     "hexdoc_item",
     "hexdoc_localize",
+    "hexdoc_smart_var",
     "hexdoc_texture",
     "hexdoc_wrap",
 ]
 
 from .extensions import IncludeRawExtension
-from .filters import hexdoc_item, hexdoc_localize, hexdoc_texture, hexdoc_wrap
+from .filters import (
+    hexdoc_item,
+    hexdoc_localize,
+    hexdoc_smart_var,
+    hexdoc_texture,
+    hexdoc_wrap,
+)